+++ R/zzz.R
@@ -1,3 +1,6 @@
 
-utils::globalVariables(c("rn", "variable", "value", "text"))
+utils::globalVariables(
+  c("rn", "variable", "value", "text", "Mode", "Name", "Stats", 
+    "area", "day", "getGeographicTrimming", "getIdCols", "link", "time", "timeId",
+    "week1", "week2", "year1", "year2"))